+++ nylas/client/restful_models.py
@@ -5,7 +5,7 @@ from enum import Enum
 from six import StringIO
 from nylas.client.restful_model_collection import RestfulModelCollection
 from nylas.client.errors import FileUploadError, UnSyncedError, NylasApiError
-from nylas.utils import timestamp_from_dt
+from nylas.utils import timestamp_from_dt, AuthMethod
 
 # pylint: disable=attribute-defined-outside-init
 
@@ -35,6 +35,7 @@ class RestfulModel(dict):
     datetime_filter_attrs = {}
     typed_dict_attrs = {}
     read_only_attrs = {}
+    auth_method = AuthMethod.BEARER
     # The Nylas API holds most objects for an account directly under '/',
     # but some of them are under '/a' (mostly the account-management
     # and billing code). api_root is a tiny metaprogramming hack to let
@@ -158,17 +159,22 @@ class NylasAPIObject(RestfulModel):
 
     def save(self, **kwargs):
         if self.id:
-            new_obj = self.api._update_resource(
-                self.cls, self.id, self.as_json(), **kwargs
-            )
+            new_obj = self._update_resource(**kwargs)
         else:
-            new_obj = self.api._create_resource(self.cls, self.as_json(), **kwargs)
-        for attr in self.cls.attrs:
-            if hasattr(new_obj, attr):
-                setattr(self, attr, getattr(new_obj, attr))
+            new_obj = self._create_resource(**kwargs)
+        self._update_values(new_obj)
 
     def update(self):
-        new_obj = self.api._update_resource(self.cls, self.id, self.as_json())
+        new_obj = self._update_resource()
+        self._update_values(new_obj)
+
+    def _create_resource(self, **kwargs):
+        return self.api._create_resource(self.cls, self.as_json(), **kwargs)
+
+    def _update_resource(self, **kwargs):
+        return self.api._update_resource(self.cls, self.id, self.as_json(), **kwargs)
+
+    def _update_values(self, new_obj):
         for attr in self.cls.attrs:
             if hasattr(new_obj, attr):
                 setattr(self, attr, getattr(new_obj, attr))